Refactor order business logic from router to service
This commit moves the business logic for order management from
`src/app/features/orders/router.py` to a new `src/app/features/orders/service.py`.

Key changes include:
- A new `OrderService` class in `service.py` now encapsulates all core operations:
    - `_to_order_public_schema`: Helper to convert Order model to public schema.
    - `create_new_order`: Handles creation of new orders, including inventory checks and transaction management.
    - `get_all_orders`: Fetches and filters a list of orders.
    - `get_order_by_public_id`: Retrieves a single order by its public ID with authorization.
    - `ship_existing_order`: Manages shipping an order.
    - `cancel_existing_order`: Handles order cancellation and stock replenishment.
- The `router.py` has been updated to delegate all business logic calls to the `OrderService`.
- Imports in both files have been cleaned up to reflect the new structure.

This refactoring improves separation of concerns, making the codebase more modular, maintainable, and testable. The router layer is now responsible for HTTP request/response handling and routing, while the service layer contains the business logic.
